fix: rename IceSAT2_startup => config
