+++ src/icesat2_tracks/analysis_db/B02_make_spectra_gFT.py
@@ -26,7 +26,7 @@ import icesat2_tracks.ICEsat2_SI_tools.spectral_estimates as spec
 import icesat2_tracks.local_modules.m_general_ph3 as M
 import icesat2_tracks.local_modules.m_spectrum_ph3 as spicke_remover
 import icesat2_tracks.local_modules.m_tools_ph3 as MT
-from icesat2_tracks.config.IceSAT2_startup import mconfig
+from icesat2_tracks.config.config import mconfig
 
 from icesat2_tracks.clitools import (
     echo,
