+++ src/icesat2_tracks/analysis_db/B05_define_angle.py
@@ -9,7 +9,7 @@ import matplotlib
 from matplotlib import pyplot as plt
 from typer import Option
 
-from icesat2_tracks.config.IceSAT2_startup import (
+from icesat2_tracks.config.config import (
     mconfig,
     color_schemes,
     font_for_print,
